lalsimutils: Zero-pad FD waveforms up to Nyquist
Now that the inner products are computed with vectorized routines,
they expect the input arrays to have a standard length up to Nyquist.
However, the LAL routines for TaylorF2 only fill their output array to
the ISCO freq. This patch ensures that when hoff(P) is called for FD
approximants, they will zero-pad up to the appropriate length.

This allows TaylorF2 to be used in inner products once again.