mcsampler.py: Fix to pinning procedure
If a value was marked as pinned, the corresponding _pdf_norm value was
not reset to one. In the case where normalized PDFs are used, this is
not an issue, since it would have been one anyway. However, unnormalized
PDFs are used for the distance prior, and such, were causing erroneous
values of the sampling prior.