mcsampler.py: Save joint prior and sampling prior values
In the same way that the integrand is cached, now the joint prior and
sampling prior are also preserved if the save_integrand option is
supplied to the integrator.

NOTE: The prior values should not be made part of the integration loop.
At some point, the MCSampler class needs to be subclassed to handle this
particular integral and the conveniences which have been stuck on it.