coalesce.sh: Preprocessing for ligolw_sqlite
Our coalescing jobs require some way of knowing the input XML files to
convert to sqlite. Due to limitations on the way we use condor, this
needs to be accomplished with a PRE script on the coalesce node. This
script is run with the condor JOBID as an argument to create that cache
before the ligolw_sqlite command is run.