lalsimutils.py: Frame reading functions can window
Added a window_shape parameter to frame_data_to_blah functions
which convert a frame file to Time/FrequencySeries. The default value 0
means no windowing is done. Otherwise, a Tukey window is applied. The data is
smoothly tapered from zero to full strength at the outer edges of the data, with
window_shape controlling over what fraction of the data the taper is done.