Added util_make_zero_noise_inj_frame
This program generates a frame with a set of injection channels from a sim inspiral table. Since things like tidal parameters are not encodable (currenly) in sim_inspiral tables, this program provides the functionality to modify the ChooseWaveformParams structure from the command line. This will allow one to set values for parameters present in the structure but not the XML table.