Complete implementation of spherical harmonics on GPU
Now supports spherical harmonics all the way up to `l_max = 8`.

Added argument to spherical harmonics function for `l_max`, so
pre-computed quantities are only generated if they might be used.
E.g., `cos(6*theta)` appears in the `l = 8` term, but none of the
others, so it is only pre-computed if `l_max >= 8`.